# SteticOpenApi.yaml
openapi: 3.0.1
info:
  title: Stetic API
  description: API para la gestión de un centro de estética
  version: 1.0.0
servers:
  - url: 'https://tiendasphy.000webhostapp.com/index.php'
...
paths:
  /clients:
    get:
      tags:
        - clients
      summary: Lista todos los clientes o filtra por nombre, dni o ciudad
      parameters:
        - name: firstname
          in: query
          description: Nombre del cliente
          required: false
          schema:
            type: string
        - name: dni
          in: query
          description: DNI del cliente
          required: false
          schema:
            type: string
        - name: city
          in: query
          description: ciudad del cliente
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Devuelve la lista completa de clientes o el cliente filtrado
          content:
            application/json:
              example:

                - id: 1
                  firstname: Juan
                  lastname: Pérez
                  DNI: "12345678A"
                  city: Barcelona
                  street: Calle Principal
                  numHouse: 123
                  age: 30
                - id: 2
                  firstname: María
                  lastname: López
                  DNI: "87654321B"
                  city: Madrid
                  street: Calle Secundaria
                  numHouse: 456
                  age: 25

    post:
      tags:
        - clients
      summary: Añadir un nuevo cliente
      requestBody:
        description: Cliente a añadir
        required: true
        content:
          application/json:
            examples:
              Example1:
                summary: Ejemplo de nuevo cliente1
                value: |
                  {
                      "firstname": "Manuel",
                      "lastname": "Martinez",
                      "DNI": "98765432C",
                      "city": "Sevilla",
                      "street": "Calle Nueva",
                      "numHouse": 789,
                      "age": 28
                  }
              Example2:
                summary: Ejemplo de nuevo cliente2
                value: |
                  {
                      "firstname": "Antonio",
                      "lastname": "Carrasco",
                      "DNI": "87878767R",
                      "city": "Pamplona",
                      "street": "Calle Nueva",
                      "numHouse": 79,
                      "age": 35
                  }

      responses:
        '201':
          description: Cliente añadido correctamente
          content:
            application/json:
              examples:
                Example1:
                  summary: cliente1 creado
                  value: |
                    {
                        "id": 1,
                        "firstname": "Manuel",
                        "lastname": "Martinez",
                        "DNI": "98765432C",
                        "city": "Sevilla",
                        "street": "Calle Nueva",
                        "numHouse": 789,
                        "age": 28
                    }
                Example2:
                  summary: cliente2 creado
                  value: |
                    {
                        "id": 2,
                        "firstname": "Antonio",
                        "lastname": "Carrasco",
                        "DNI": "87878767R",
                        "city": "Pamplona",
                        "street": "Calle Nueva",
                        "numHouse": 79,
                        "age": 35
                    }
        '400':
          description: Error de validación en los datos del cliente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
  /clients/{clientId}:
    get:
      tags:
        - clients
      summary: Obtiene un cliente por ID
      parameters:
        - name: clientId
          in: path
          description: ID del cliente
          required: true
          schema:
            type: integer
            format: int64
            example: 1

      responses:
        '200':
          description: Devuelve el cliente solicitado
          content:
            application/json:
              examples:
                Example1:
                  summary: cliente con ID 1
                  value: |
                    {
                        "id": 1,
                        "firstname": "Juan",
                        "lastname": "Pérez",
                        "DNI": "12345678A",
                        "city": "Barcelona",
                        "street": "Calle Principal",
                        "numHouse": 123,
                        "age": 30
                    }
                Example2:
                  summary: cliente con ID 2
                  value: |
                    {
                        "id": 2,
                        "firstname": "Antonio",
                        "lastname": "Carrasco",
                        "DNI": "87878767R",
                        "city": "Pamplona",
                        "street": "Calle Nueva",
                        "numHouse": 79,
                        "age": 35
                    }
        '404':
          description: Cliente no encontrado
          content:
            application/json:
              example:
                code: 404
                message: "Cliente no encontrado"
  /client/{DNI}:
    get:
      tags:
        - clients
      summary: Obtiene un cliente por DNI
      parameters:
        - name: DNI
          in: path
          description: DNI del cliente
          required: true
          schema:
            type: string
            example: Ejemplo con el DNI 87878767R



      responses:
        '200':
          description: Devuelve el cliente solicitado
          content:
            application/json:
              examples:
                Example1:
                  summary: cliente con DNI 87878767R
                  value: |
                    {
                        "id": 2,
                        "firstname": "Antonio",
                        "lastname": "Carrasco",
                        "DNI": "87878767R",
                        "city": "Pamplona",
                        "street": "Calle Nueva",
                        "numHouse": 79,
                        "age": 35
                    }
                Example2:
                  summary: cliente con DNI 12345678A
                  value: |
                    {
                        "id": 1,
                        "firstname": "Juan",
                        "lastname": "Pérez",
                        "DNI": "12345678A",
                        "city": "Barcelona",
                        "street": "Calle Principal",
                        "numHouse": 123,
                        "age": 30
                    }

        '404':
          description: Cliente no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
    delete:
      tags:
        - clients
      summary: Elimina un cliente por DNI
      parameters:
        - name: DNI
          in: path
          description: DNI del cliente
          required: true
          schema:
            type: string
      responses:
        '204':
          description: El cliente se ha eliminado correctamente
        '404':
          description: Cliente no encontrado
          content:
            application/json:
              example:
                code: 404
                message: "Cliente no encontrado"
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
    put:
      tags:
        - clients
      summary: Modifica un cliente por DNI
      parameters:
        - name: DNI
          in: path
          description: DNI del cliente
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            examples:
              Example1:
                summary: Modificar cliente con DNI 87878767R
                value: |
                  {
                      "firstname": "Antonio",
                      "lastname": "Carmona",
                      "city": "Palencia",
                      "street": "Calle Vieja",
                      "numHouse": 79,
                      "age": 35
                  }
              Example2:
                summary: Modificar cliente con DNI 12345678A
                value: |
                  {
                      "firstname": "Santiago",
                      "lastname": "Perez",
                      "city": "Vigo",
                      "street": "Carretera",
                      "numHouse": 79,
                      "age": 45
                  }
      responses:
        '200':
          description: El cliente se ha modificado correctamente
        '404':
          description: Cliente no encontrado
          content:
            application/json:
              example:
                code: 404
                message: "Cliente no encontrado"
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
        errors:
          type: array
          items:
            type: string
    InternalServerError:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
    BadRequest:
      type: object
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
      example:
        code: 400
        message: Validation error
        errors:
          - El nombre es obligatorio
          - El precio debe ser mayor que cero